requirements.txt
openai>=1.23.0
geopandas>=0.14.0
rasterio>=1.3.0
shapely>=2.0
langchain>=0.1
PyYAML>=6.0
python-dotenv>=1.0
pandas>=2.0
sentence-transformers
tiktoken
faiss-cpu
fastapi
uvicorn[standard]
pydantic
httpx
python-jose[cryptography]
passlib[bcrypt]
slowapi
loguru
streamlit
plotly
graphviz
pytest>=7.0
pytest-asyncio
networkx